// .kilo/worktrees/amplified-cave/src/api/teaching/types.ts
export type EnterScoreBody = {
  student_id: number
  subject_id: number
  session_id: number
  semester_id: number
  ca: number
  exam: number
}